requirements.txt
qdrant-client
sentence-transformers
flask
pyyaml
requests